Added login handler
